% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_get_product_export.R
\name{ca_get_product_export}
\alias{ca_get_product_export}
\title{Retreive products via file export (API)}
\usage{
ca_get_product_export(account = "US")
}
\arguments{
\item{account}{target account/profile}
}
\description{
Retreive products via file export (API)
}
