% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_get_products.R
\name{ca_get_products}
\alias{ca_get_products}
\title{Return product info (possibly by SKU)}
\usage{
ca_get_products(sku = NULL,
  access_token = Sys.getenv("CA_ACCESS_TOKEN"))
}
\arguments{
\item{sku}{product SKU}

\item{access_token}{dev access token}
}
\description{
Return product info (possibly by SKU)
}
