% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_parse_datetime.R
\name{ca_parse_datetime}
\alias{ca_parse_datetime}
\title{Parse datetime output from ChannelAdvisor API}
\usage{
ca_parse_datetime(x)
}
\arguments{
\item{x}{datetime as character}
}
\description{
Parse datetime output from ChannelAdvisor API
}
