% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_parse_response.R
\name{ca_parse_response}
\alias{ca_parse_response}
\title{Parse a response object returned from the ChannelAdvisor API}
\usage{
ca_parse_response(r)
}
\arguments{
\item{x}{response content}
}
\description{
Parse a response object returned from the ChannelAdvisor API
}
