% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_set_account.R
\name{ca_set_account}
\alias{ca_set_account}
\title{Set the current ChannelAdvisor account profile (i.e., locale)}
\usage{
ca_set_account(account = NULL)
}
\arguments{
\item{account}{account name}
}
\description{
Set the current ChannelAdvisor account profile (i.e., locale)
}
