% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_ship_order.R
\name{ca_ship_order}
\alias{ca_ship_order}
\title{Mark order as shipped and update tracking}
\usage{
ca_ship_order(orderid, tracking, carrier, class)
}
\arguments{
\item{orderid}{order ID}

\item{tracking}{tracking number}

\item{carrier}{shipping carrier code}

\item{class}{shipping class code}
}
\description{
Mark order as shipped and update tracking
}
