% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_update_product.R
\name{ca_update_product}
\alias{ca_update_product}
\title{Update product info}
\usage{
ca_update_product(id = NULL, sku = NULL, data = NULL, account = "US")
}
\arguments{
\item{id}{product ID}

\item{sku}{product SKU}

\item{data}{product information (list/tibble)}

\item{account}{target account/profile}
}
\description{
Update product info
}
