% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_upload_products.R
\name{ca_upload_products}
\alias{ca_upload_products}
\title{Upload a product/inventory file}
\usage{
ca_upload_products(file, type = c("AddUpdate", "UpdateOnly", "AddOnly"))
}
\arguments{
\item{file}{full path to file to upload}

\item{type}{import type}
}
\description{
Upload a product/inventory file
}
