% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_api.R
\name{ca_api}
\alias{ca_api}
\title{Make a call to the ChannelAdvisor API}
\usage{
ca_api(endpoint = "Products", account = "US", ...)
}
\arguments{
\item{endpoint}{API endpoint}

\item{account}{account/profile}

\item{query}{named list with query to submit}
}
\description{
Make a call to the ChannelAdvisor API
}
