% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_upload_products.R
\name{ca_upload_products}
\alias{ca_upload_products}
\title{Upload a product/inventory file}
\usage{
ca_upload_products(
  file = NULL,
  data = NULL,
  type = c("AddUpdate", "UpdateOnly", "AddOnly"),
  account = Sys.getenv("CHANNELADVISOR_ACCOUNT")
)
}
\arguments{
\item{file}{full path to file to upload}

\item{data}{data.frame/tibble to upload}

\item{type}{import type}

\item{account}{account/profile}
}
\description{
Upload a product/inventory file
}
